+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Fixed
 
 - `Scroller.Scroller` to wheel normally when `focusableScrollbar` prop is `byEnter`.
+- `sandstone/ContextualPopupDecorator` to correctly manage focus when changing its open state
 
 ## [1.0.0-alpha.2] - 2020-03-03